fix(frontend): fix following
	-add line endings
	-move files with dummy data to a folder js/data to make it easy to change these files later
