+++ frontend/src/css/addtalk.scss
