+++ frontend/src/css/lightningtalk.scss
