+++ frontend/src/js/data/lightningtalks.js
