+++ frontend/src/js/lightningtalk.js
