+++ frontend/src/js/register.js
