+++ frontend/src/views/public/Nav.svelte
